selectCandidateKeypoints: enhancements and options
Modify the interface of selectCandidateKeypoints: instead of a
set of keypoints, pass the whole image and do kps detection and
selection in place. Add option to cap the number of keypoints:
if the number of detected keypoints is larger than the requested
one, uniformly select a subset of them. Add options to tweak the
parameters of the keypoints detector. Update VisualOdometry
and the related test scripts to comply with the new interface.